Revert "Task 4 (#5)"
This reverts commit 3d22ffdf7e2e1ee5dc5e39f9fec37ff7c5b88dc0.
