+++ .circleci/config.yml
@@ -3,7 +3,7 @@ version: 2
 jobs:
   build:
     docker:
-      - image: circleci/python:3.7.0-browsers
+      - image: circleci/python:3.6.6-browsers
       
     working_directory: ~/repo
 
@@ -16,23 +16,14 @@ jobs:
             python3 -m venv venv
             . venv/bin/activate
             pip install -r requirements.txt
-            curl -L  https://qameta.bintray.com/generic/io/qameta/allure/allure/2.7.0/allure-2.7.0.zip -o allure.zip
-            unzip ./allure.zip
 
       - run:
           name: run tests
           command: |
             . venv/bin/activate
-            mkdir ./reports
-            python3 -m pytest -v -n 4 tests --alluredir ./reports
-          when: always
-
-      - run:
-          name: generate reports
-          command: |
-            allure-2.7.0/bin/allure generate -c ./reports
-          when: always
+            cd ./tests/api_tests
+            pytest -n 4 -v .
 
       - store_artifacts:
-          path: ./allure-report
-          destination: allure-report
+          path: test-reports
+          destination: test-reports
