--- config.ini
@@ -1,14 +1,0 @@
-[BASE]
-timeout = 20
-
-[TEST]
-browser = chrome
-headless = True
-fullscreen = False
-browser_resolution = HD
-
-[ENV]
-jira_base_url = http://jira.hillel.it:8080
-jira_api_url = /rest/api/2
-username = Alexander_Dovgalenko
-password = test_pass
