--- test_data/csv/create_issue_negative.csv
@@ -1,2 +1,0 @@
-AQAPYTHON|AQAPYTHON -- {} -- summary 258 -- sjhdbvhsdKASVDHVAvtdf fdsvfdfb sdfbsfd bdyfeigfbsdveiudsfbdsfbdfshHJdsfbdsVAB BNsjhdbvhsdbaghvaghvasdjf adf gadsv gdsag hadsgsadkjh gdhsajkgasdg hsadg hjksdag fhadgjf gsadhjf fadsvbmns bsajckhg wdkvshjCKASVDHVASHGSADBVdgC Vgvtu -- summary 258|This issue have been created via Jira UI for testing purposes in terms of AQAPython project|Bug|Summary must be less than 255 characters.
-AQAPYTHON||This issue have been created via Jira UI for testing purposes in terms of AQAPython project|Bug|You must specify a summary of the issue.
