--- test_data/csv/create_issue_positive.csv
@@ -1,1 +1,0 @@
-AQAPYTHON|AQAPYTHON -- {} -- UI test issue by AlexDovgalenko|This issue have been created via Jira REST API for testing purposes in terms of AqAPython project|Bug
