--- test_data/csv/update_issue.csv
@@ -1,3 +1,0 @@
-summary|AQAPYTHON --  API test issue by AlexDovgalenko -- U-P-D-A-T-E-D
-severity|Blocker
-assignee|Alexander_Dovgalenko
