--- test_data/jsons/create_issue
@@ -1,1 +1,0 @@
-{"fields": {"project":{"key": "AQAPYTHON"},"summary": "AQAPYTHON -- %s -- API test issue by AlexDovgalenko","description": "This issue have been created via Jira REST API for testing purposes in terms of AqAPython project","issuetype": {"name": "Bug"}}}
