+++ tests/api_tests/test_issue_manipulations.py
@@ -1,18 +1,16 @@
 import requests
 from configurations import config
 from utils import common_utils
-import logging
+# import logging
 import pytest
 import os.path
-import allure
-
-step = allure.step
 
 
-@allure.story("Testing Jira issues manipulations via API")
 class TestIssueManipulations:
 
+    os.path.dirname('../../__file__')
 
+    # logger = logging.getLogger()
     baseUrl = config.Config.JIRA_BASE_URL
     api_url = config.Config.JIRA_REST_API_URL
     username = config.Config.JIRA_USERNAME
@@ -25,7 +23,7 @@ class TestIssueManipulations:
 
     @pytest.fixture(scope='session', autouse=True)
     def payload_type(payload_type):
-        my_path = os.path.abspath(os.path.dirname('__file__'))
+        my_path = os.path.abspath(os.path.dirname('../../__file__'))
         payload = []
         if payload_type == 'create':
             path = os.path.join(my_path, "test_data/jsons/create_issue_payloads")
@@ -43,7 +41,6 @@ class TestIssueManipulations:
             requests.request("DELETE", self.baseUrl+self.api_url+"/issue/"+self.issue_id[x], headers=self.headers, auth=(self.username, self.password))
             print("deleting issue with ID: {}".format(self.issue_id[x]))
 
-    @allure.title("Create Jira issue via API")
     @pytest.mark.parametrize("payload, expected_status_code", payload_type("create"))
     def test_create_issue(self, payload, expected_status_code):
         datetime = common_utils.get_current_datetime_str()
@@ -54,14 +51,12 @@ class TestIssueManipulations:
             global issue_id
             self.issue_id.append(r.json()['id'])
 
-    @allure.title("Update Jira issue via API")
     @pytest.mark.parametrize("payload, expected_status_code", payload_type("update"))
     def test_update_issue(self, payload, expected_status_code, test_create_issue_api):
         issue_id = test_create_issue_api
         r = requests.request("PUT", self.baseUrl+self.api_url+'/issue/'+issue_id, data=payload, headers=self.headers, auth=(self.username, self.password))
         assert r.status_code == int(expected_status_code), "Status code {0} is no match to the expected Status code {1} on attempt of updating issue".format(r.status_code, int(expected_status_code))
 
-    @allure.title("Search for issues in Jira via API")
     @pytest.mark.parametrize("payload, expected_status_code", payload_type("search"))
     def test_search_for_issue(self, payload, expected_status_code):
         r = requests.request("POST", self.baseUrl+self.api_url+'/search', data=payload, headers=self.headers, auth=(self.username, self.password))
